:recycle:refactor: Change MyTab Common Component
