+++ src/app/mypage/layout.tsx
@@ -1,7 +1,6 @@
 import AuthServiceHeader from '@/components/layout/AuthServiceHeader'
 import BottomNavigation from '@/components/layout/BottomNavigation'
 import Footer from '@/components/layout/Footer'
-import MyInfoLookUp from '@/components/mypage/myInfoLookUp'
 
 export default function Layout({
   children,
